chore: Move individual UI components into folders
Doing so to aid in isolating components with their tests, helps give better boundaries between components.
