+++ src/components/ui/Provider/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ChakraProvider, defaultSystem, Theme } from '@chakra-ui/react';
-import { ColorModeProvider, type ColorModeProviderProps } from './color-mode';
+import { ColorModeProvider, type ColorModeProviderProps } from '../ColorMode';
 
 export function Provider(props: ColorModeProviderProps) {
     return (
